+++ schittkowski/problem_49.cc
@@ -178,7 +178,8 @@ BOOST_AUTO_TEST_CASE (schittkowski_problem49)
   solver_t::problem_t::intervals_t intervals;
   intervals.push_back (G<functionType_t>::makeInterval (7., 7.));
   intervals.push_back (G<functionType_t>::makeInterval (6., 6.));
-  solver_t::problem_t::scales_t scales (g->outputSize (), 1.);
+  solver_t::problem_t::scales_t scales
+    (static_cast<std::size_t> (g->outputSize ()), 1.);
 
   problem.addConstraint (g, intervals, scales);
